Adds function for find UUID
get_UUID function added to find disk UUID based on mount
point. This will be used in a simple GUI for set up of
backups. User will select a drive; i.e., mount point. This
function will get the UUID and that will be saved in the
config.json file. Then, when we got to backup, we can look for
drive by UUID and be able to backup up no matter where mounted

Updates to get rsync_backup.py working, more testing needed.
  - Should make a 'parent' directory for backups like
    Time Machine:
       Backups.backupsdb/
           computer name/
                 backups