Minimal settings GUI and config updates
The settings GUI is 'working', but need to add more features
such as enabling automatic backups. Does warn if tries to enable
options or backup if NOT root, as must be root to get full
system backup with rsync.

Must figure out issue that appear in rsync test; not sure what is
Also, come up with regex for better progress parsing