Enables settings restore on update
When the package is installed/updated, a check to see if an
existing installation is run. If existing install found, the
config.json file is copied to the users home directory. Pip will
then uninstall the package and then install the new version. After
new install, method is run to check for the backed up config.json
file. If the file is found the data from that and the new config
file are read in. Data from all tags in the backup config file are
copied to the new config file. Thus, all previous data are
maintained.