Attempt to fix backup hang bug
I noticed that some of the back up instances were hanging when
running on cron. I assumed this was because when a lock file
was encountered, the program stopped, but the lock was deleted.
Now, the lock file is only deleted after backup is done. This
may fix the problem, but will have to test.

Also fixed exclusions so that backup disk mount point is
excluded an not just the backup directory