Add signal catching/cancel
Using the signal package, added catching of important signals
to run the cancel() method to stop the backup. If this happens,
NONE of the cleanup is performed.

May want to change this so that 'Latest' link is NOT delete while
the backup is in progress and is only changes after backup is
finished. Could also simply move failed backups to hidden
directory of same name. Then on future run, use the most recent
hidden .inprogress directory for backup, that is, move that
directory to the current backup date, then start the backup
on that. This may save some time.