Updates to use partial backups as dst
Partial backups are left when the backup is cancelled; this is
done to reduce cancel time. However, on the next backup, if
any partial backups, then the lastest partial backup is used as
destination directory; this should speed up the backup time.

Also created a __removeLock() method that is called before
return statements to make sure that the lock file is removed
correctly