Fixes logging file and config usage
When trying to log to /var/log, ran into some issues, so
now logging to home directory of root user.

Error in implimintation of the config.json data, namely, in the
directories to exclude. This has now been fixed.

Also added .rstrip() to .readline() during transfer, with check
of (line != '') for when text stops coming from rsync.

v0.0.19