Fixes unfinished updates that caused crash
In previous version (or 2), made some big changes to code to try
to improve fault tolerance and move backups to threads to keep
GUI responsive. However, failed to finish creating methods for one
of the threads.

There is now a _monitorStatus method that is run as thread and
emits Qt signals to update the progress bar and other text
properities. A Qt signal has also been added to update the
'Days since last backup' text after backup finishes (or fails).

Adds rsyncStats attribute to the rsyncBackup class that is set
to zero (0, good) or one (1, bad) based on return code from rsync.
This is done so that the GUI knows how to update the 'Days since last
backup' text.

Fixes bug in rsyncBackup where the time module was not imported.

v0.0.22