Bug fix, updates install script
Bug fix corrects a missing variable; variable had been changed
to class attribute at some point and was not updated in one part
of code.

Install script now uses the version.py file to get the version
number. Also, in the _post_install method there is check for
if install_dir is None. This way the script does not break on
initial install.

v0.1.0